+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-7571 - solved performance issues - bulk QR-code generation 
 - SC-6294 - Introduce Typescript in schulcloud-server
 - SC-7028 - Add Course Component API Specification document
+- SC-7476 - Prevent hash generation if user has account
 
 ### Changed